Move ScanSpec to utils package (it is not part of public API)
