+++ src/main/java/io/github/classgraph/ClassInfoUnlinked.java
@@ -36,6 +36,7 @@ import java.util.Map;
 import io.github.classgraph.utils.Join;
 import io.github.classgraph.utils.LogNode;
 import io.github.classgraph.utils.Parser.ParseException;
+import io.github.classgraph.utils.ScanSpec;
 
 /**
  * Class information that has been directly read from the binary classfile, before it is cross-linked with other
