+++ src/main/java/io/github/classgraph/classloaderhandler/EquinoxContextFinderClassLoaderHandler.java
@@ -28,10 +28,10 @@
  */
 package io.github.classgraph.classloaderhandler;
 
-import io.github.classgraph.ScanSpec;
 import io.github.classgraph.utils.ClasspathOrder;
 import io.github.classgraph.utils.LogNode;
 import io.github.classgraph.utils.ReflectionUtils;
+import io.github.classgraph.utils.ScanSpec;
 
 /** Extract classpath entries from the Eclipse Equinox ContextFinder ClassLoader. */
 public class EquinoxContextFinderClassLoaderHandler implements ClassLoaderHandler {
