+++ src/main/java/io/github/classgraph/classloaderhandler/OSGiDefaultClassLoaderHandler.java
@@ -30,10 +30,10 @@ package io.github.classgraph.classloaderhandler;
 
 import java.io.File;
 
-import io.github.classgraph.ScanSpec;
 import io.github.classgraph.utils.ClasspathOrder;
 import io.github.classgraph.utils.LogNode;
 import io.github.classgraph.utils.ReflectionUtils;
+import io.github.classgraph.utils.ScanSpec;
 
 /**
  * Handle the OSGi DefaultClassLoader.
