+++ src/main/java/io/github/classgraph/utils/JarUtils.java
@@ -43,8 +43,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import io.github.classgraph.ScanSpec;
-
 /**
  * Jarfile utilities.
  */
